+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix compact folders when grouping by tag
 - Fix description of 'hide tree when empty' option
+- Add support for total todo count in view title
 
 ## v0.0.195 - 2021-01-20